Don’t attempt to guarantee that the interval will continue to run if `func` throws
Because, in Node, an uncaught exception will probably cause the process to
exit.

The problem preventing us testing this in Node (see the deleted test code)
turned out to be https://github.com/facebook/jest/issues/5620, although it
didn’t make sense to test that in Node after all given that the process would
exit—the interval wouldn’t continue to run regardless of how this library was
written.

We could test this in a browser environment instead, but probably better for
this library to not try to guarantee that the interval will run, as is now
documented in the code and in the README.